+++ src/Aks/Aks/Commands/SetAzureRmAks.cs
@@ -278,6 +278,14 @@ namespace Microsoft.Azure.Commands.Aks
                                     defaultAgentPoolProfile.NodeLabels.Add(key.ToString(), NodePoolLabel[key].ToString());
                                 }
                             }
+                            if (this.IsParameterBound(c => c.NodePoolTags))
+                            {
+                                defaultAgentPoolProfile.Tags = new Dictionary<string, string>();
+                                foreach (var key in NodePoolTags.Keys)
+                                {
+                                    defaultAgentPoolProfile.Tags.Add(key.ToString(), NodePoolTags[key].ToString());
+                                }
+                            }
                         }
 
                         if (this.IsParameterBound(c => c.KubernetesVersion) && this.IsParameterBound(c => c.NodeImageOnly))
@@ -450,7 +458,8 @@ namespace Microsoft.Azure.Commands.Aks
             return this.IsParameterBound(c => c.NodeCount) || this.IsParameterBound(c => c.NodeOsDiskSize) ||
                 this.IsParameterBound(c => c.NodeVmSize) || this.IsParameterBound(c => c.EnableNodeAutoScaling) ||
                 this.IsParameterBound(c => c.NodeMinCount) || this.IsParameterBound(c => c.NodeMaxCount) || 
-                this.IsParameterBound(c => c.NodePoolMode) || this.IsParameterBound(c => c.NodePoolLabel);
+                this.IsParameterBound(c => c.NodePoolMode) || this.IsParameterBound(c => c.NodePoolLabel) ||
+                this.IsParameterBound(c => c.NodePoolTags);
         }
     }
 }